Refactor S3 usage and always apply protocol fix
