+++ backend/src/graphql/resolvers/images/images.ts
@@ -1,4 +1,5 @@
 import CONFIG, { isS3configured } from '@config/index'
+import type { FileDeleteCallback, FileUploadCallback } from '@src/uploads/types'
 
 import { images as imagesLocal } from './imagesLocal'
 import { images as imagesS3 } from './imagesS3'
@@ -6,11 +7,6 @@ import { images as imagesS3 } from './imagesS3'
 import type { FileUpload } from 'graphql-upload'
 import type { Transaction } from 'neo4j-driver'
 
-export type FileDeleteCallback = (url: string) => Promise<void>
-
-export type FileUploadCallback = (
-  upload: Pick<FileUpload, 'createReadStream' | 'mimetype'> & { uniqueFilename: string },
-) => Promise<string>
 export interface DeleteImageOpts {
   transaction?: Transaction
   deleteCallback?: FileDeleteCallback
