+++ backend/src/graphql/resolvers/images/imagesS3.ts
@@ -1,34 +1,23 @@
 import path from 'node:path'
 
-import { S3Client, DeleteObjectCommand, ObjectCannedACL } from '@aws-sdk/client-s3'
-import { Upload } from '@aws-sdk/lib-storage'
 import { UserInputError } from 'apollo-server'
+import { FileUpload } from 'graphql-upload'
 import slug from 'slug'
 import { v4 as uuid } from 'uuid'
 
-import { S3Configured } from '@config/index'
+import type { S3Configured } from '@config/index'
+import { s3Service } from '@src/uploads/s3Service'
+import { FileUploadCallback } from '@src/uploads/types'
 
 import { wrapTransaction } from './wrapTransaction'
 
-import type { Image, Images, FileDeleteCallback, FileUploadCallback } from './images'
-import type { FileUpload } from 'graphql-upload'
+import type { Image, Images } from './images'
 
 export const images = (config: S3Configured) => {
-  // const widths = [34, 160, 320, 640, 1024]
-  const { AWS_BUCKET: Bucket, S3_PUBLIC_GATEWAY } = config
-
-  const { AWS_ENDPOINT, AWS_ACCESS_KEY_ID, AWS_SECRET_ACCESS_KEY } = config
-  const s3 = new S3Client({
-    credentials: {
-      accessKeyId: AWS_ACCESS_KEY_ID,
-      secretAccessKey: AWS_SECRET_ACCESS_KEY,
-    },
-    endpoint: AWS_ENDPOINT,
-    forcePathStyle: true,
-  })
+  const s3 = s3Service(config, 'original')
 
   const deleteImage: Images['deleteImage'] = async (resource, relationshipType, opts = {}) => {
-    const { transaction, deleteCallback = s3Delete } = opts
+    const { transaction, deleteCallback = s3.deleteFile } = opts
     if (!transaction) return wrapTransaction(deleteImage, [resource, relationshipType], opts)
     const txResult = await transaction.run(
       `
@@ -58,7 +47,7 @@ export const images = (config: S3Configured) => {
   ) => {
     if (typeof imageInput === 'undefined') return
     if (imageInput === null) return deleteImage(resource, relationshipType, opts)
-    const { transaction, uploadCallback, deleteCallback = s3Delete } = opts
+    const { transaction, uploadCallback, deleteCallback = s3.deleteFile } = opts
     if (!transaction)
       return wrapTransaction(mergeImage, [resource, relationshipType, imageInput], opts)
 
@@ -95,51 +84,13 @@ export const images = (config: S3Configured) => {
 
   const uploadImageFile = async (
     uploadPromise: Promise<FileUpload> | undefined,
-    uploadCallback: FileUploadCallback | undefined = s3Upload,
+    uploadCallback: FileUploadCallback | undefined = s3.uploadFile,
   ) => {
     if (!uploadPromise) return undefined
     const upload = await uploadPromise
     const { name, ext } = path.parse(upload.filename)
     const uniqueFilename = `${uuid()}-${slug(name)}${ext}`
-    const Location = await uploadCallback({ ...upload, uniqueFilename })
-    if (!S3_PUBLIC_GATEWAY) {
-      return Location
-    }
-    const publicLocation = new URL(S3_PUBLIC_GATEWAY)
-    publicLocation.pathname = new URL(Location).pathname
-    return publicLocation.href
-  }
-
-  const s3Upload: FileUploadCallback = async ({ createReadStream, uniqueFilename, mimetype }) => {
-    const s3Location = `original/${uniqueFilename}`
-    const params = {
-      Bucket,
-      Key: s3Location,
-      ACL: ObjectCannedACL.public_read,
-      ContentType: mimetype,
-      Body: createReadStream(),
-    }
-    const command = new Upload({ client: s3, params })
-    const data = await command.done()
-    const { Location } = data
-    if (!Location) {
-      throw new Error('File upload did not return `Location`')
-    }
-    return Location
-  }
-
-  const s3Delete: FileDeleteCallback = async (url) => {
-    let { pathname } = new URL(url, 'http://example.org') // dummy domain to avoid invalid URL error
-    pathname = pathname.substring(1) // remove first character '/'
-    const prefix = `${Bucket}/`
-    if (pathname.startsWith(prefix)) {
-      pathname = pathname.slice(prefix.length)
-    }
-    const params = {
-      Bucket,
-      Key: pathname,
-    }
-    await s3.send(new DeleteObjectCommand(params))
+    return await uploadCallback({ ...upload, uniqueFilename })
   }
 
   const images: Images = {
