+++ backend/src/uploads/types.ts
@@ -1,0 +1,7 @@
+import type { FileUpload } from 'graphql-upload'
+
+export type FileDeleteCallback = (url: string) => Promise<void>
+
+export type FileUploadCallback = (
+  upload: Pick<FileUpload, 'createReadStream' | 'mimetype'> & { uniqueFilename: string },
+) => Promise<string>
